+++ graph/grdevice.c
@@ -276,7 +276,7 @@
   {
     int  len;
 
-    len = strlen(string);
+    len = (int)strlen(string);
     grWriteCellString( &target->bitmap, x, y, string, color );
     if (target->refresh_rect)
       target->refresh_rect( target, x, y, 8*len, 8 );
